WIP: Add `scalable_tile_flags` to #iree_codegen.lowering_config
This adds a new `scalable_tile_flags` property to the lowering config
and hooks it up to tiling and the generic vectorizer. This property
has the same shape as the `tile_sizes`, but all values are just booleans
that mark if the corresponding tile size is scalable.

This is still very much a test/proof-of-concept patch.