Support scalable tile sizes in #iree_codegen.lowering_config
This extends the lowering config so that at each level tiles can be
marked as scalable using an optional array of bools. The parser/printer
has been updated to display this in the canonical MLIR syntax (i.e. the
size wrapped in square brackets).

So for example, this lowering config:

```
  #iree_codegen.lowering_config<tiling_levels
	= [[128, 128, 0], [1, [32], 0], [0, 0, 1], [0, 0, 0]]>
```

Means that the middle tile size (32) of the second tiling level is
scalable.

These are not yet used, but will be in subsequent patches.